fix putnbrbase

+++ ft_utils.c
@@ -6,14 +6,14 @@
 /*   By: mmeguedm <mmeguedm@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/20 22:59:53 by mmeguedm          #+#    #+#             */
-/*   Updated: 2022/05/22 18:04:01 by mmeguedm         ###   ########.fr       */
+/*   Updated: 2022/05/22 20:01:42 by mmeguedm         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 // #include "ft_utils.h"
 // #include <unistd.h>
 
-#include "includes.h"
+#include "ft_printf.h"
 
 int	ft_strlen(char *str)
 {
@@ -23,7 +23,6 @@ int	ft_strlen(char *str)
 	while (str[i])
 		i++;
 	return(i);
-
 }
 
 void	ft_putchar(char c)
@@ -59,7 +58,9 @@ int	ft_checkbase(char	*base)
 	}
 	return (1);
 }
-void	ft_putnbrbase(unsigned long long int nb, char *base)
+#include <stdio.h>
+
+void	ft_putnbrbase(long long int nb, char *base)
 {
 	if (!ft_checkbase(base))
 		return ;
@@ -68,11 +69,12 @@ void	ft_putnbrbase(unsigned long long int nb, char *base)
 		nb *= -1;
 		ft_putchar('-');
 	}
-	if (nb >= ft_strlen(base))
+	if (nb >= (long long int)ft_strlen(base))
 	{
 		ft_putnbrbase(nb / ft_strlen(base), base);
 		ft_putnbrbase(nb % ft_strlen(base), base);
 	}
 	else
 		ft_putchar(base[nb % ft_strlen(base)]);
+
 }
